change-12042024-part2 - Broke apart main file to force dev to use local and prod to use git.
